Move sensor(s) tests to match code under test
The test package name now matches that of the package that's being
targetted by these tests, helping to make the test structure a bit
clearer in advance of adding more tests.
