+++ tests/sensors/test_extra_commands.py
