+++ tests/sensors/test_obsdata.py
